refactor: axis scroll out of viewer
